+++ lmeval/models/gemini.py
@@ -18,7 +18,7 @@ from litellm import check_valid_key, completion, batch_completion
 from litellm import ModelResponse
 
 from ..media import Modality
-from .litellm_model import LiteLLMModel
+from .litellm import LiteLLMModel
 # try:
 #     import google.generativeai as genai
 # except ImportError: