+++ lmeval/prompts/multi_choices_prompts.py
@@ -20,6 +20,93 @@ from ..task import Task
 from .prompt import Prompt
 from ..enums import TaskType
 
+MULTI_ANSWER_TEMPLATE = """
+        Accurately answer the following question:
+
+        {{question.question}}
+
+        Choices:
+        {{question.multi_choices}}
+
+        Instructions:
+        - Carefully read the question and all options.
+        - Select all the correct answers.
+        - Respond ONLY with the letters ({{question.letters}}) corresponding to the correct answers.
+        - Do not include any explanations, additional text, or punctuation in your response.
+        - If there are multiple correct answers, separate them with a comma (,). For example "A, B, C".
+        Your answer ({{question.letters}}):
+    """
+
+
+class MultiChoicesMultiAnswersPrompt(Prompt):
+    def __init__(self,
+                template: str = MULTI_ANSWER_TEMPLATE,
+                name: str = "Multi Choices Multi Answer Picker",
+                description: str = "Ask the model to return the letters associated with potentially multiple correct answers",
+                task_type = TaskType.multiple_choices_multiple_answers,
+                url: str = '',
+                version: str = '1.0'):
+
+            super().__init__(name=name, description=description,
+                            task_type=task_type, template=template, url=url,
+                            version=version)
+    def render(self, question: Question, task: Task) -> str:
+        "Render prompt for a given question and task"
+
+        if task.type != self.task_type:
+            raise ValueError(f"Task type {task.type} does not match prompt task type {self.task_type}")
+
+        assert question.answer not in question.choices, f"Answer {question.answer} should not be in other choices. {question.choices}"
+
+
+        version = self.version_string()
+
+        if version in question.prompt_cache:
+            question.multi_choices = question.prompt_cache[version]['multi_choices']
+            question.letters = question.prompt_cache[version]['letters']
+            question.answer_letter = question.prompt_cache[version]['answer_letter']
+        else:
+            possible_answers = [question.answer] + question.additional_answers +  question.choices
+            random.shuffle(possible_answers)
+
+            # Construct the list of possible answers
+            choices_list = []
+            letters_list = []
+            correct_letters = []
+            for idx, answer in enumerate(possible_answers):
+                letter = ascii_uppercase[idx]
+                # don't put space between letter and answer it decrease accuracy...
+                choices_list.append(f"{letter}:{answer}")
+                letters_list.append(letter)
+
+                # mark the answer and additional answers as correct
+                if answer == question.answer:
+                    correct_letters.append(letter)
+                if answer in question.additional_answers:
+                    correct_letters.append(letter)
+
+            question.answer_letter = ', '.join(correct_letters)
+
+            # flatten
+            multi_choices = "\n".join(choices_list)
+            letters = ', '.join(letters_list)
+
+            # assign to the current question
+            question.multi_choices = multi_choices
+            question.letters = letters
+
+            # store assignements for reuse accross models to have the exact same question
+            question.prompt_cache[version] = {
+                'multi_choices': multi_choices,
+                'letters': letters,
+                'answer_letter': question.answer_letter
+            }
+
+        # render full template
+        template = TemplateEngine(self.template)
+        return template.render(question=question, task=task)
+
+
 TEMPLATE = """
         Accurately answer the following question:
 
@@ -36,6 +123,7 @@ TEMPLATE = """
 
         Your answer ({{question.letters}}):
     """
+
 
 class MultiChoicesPrompt(Prompt):